tests(parser): Split tests files
